% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{ate}
\alias{ate}
\title{Estimating average effect of discrete treatment}
\usage{
ate(y, a, x, nsplits=2, sl.lib=c("SL.earth","SL.gam","SL.glm","SL.glmnet",
  "SL.glm.interaction", "SL.mean","SL.ranger","rpart"))
}
\arguments{
\item{y}{outcome of interest.}

\item{a}{discrete treatment.}

\item{x}{covariate matrix.}

\item{nsplits}{integer number of sample splits for nuisance estimation.
If nsplits=1, sample splitting is not used, and nuisance functions are estimated
on full sample (in which case validity of SEs/CIs requires empirical
process conditions). Otherwise must have nsplits>1.}

\item{sl.lib}{algorithm library for SuperLearner.
Default library includes "earth", "gam", "glm", "glmnet", "glm.interaction",
"mean", "ranger", "rpart.}
}
\value{
A list containing the following components:
\item{res}{ estimates/SEs/CIs/p-values for population means and relevant contrasts.}
\item{nuis}{ subject-specific estimates of nuisance functions (i.e., propensity score and outcome regression) }
\item{ifvals}{ matrix of estimated influence function values.}
}
\description{
\code{ate} is used to estimate the mean outcome in a population had all subjects received given levels of a discrete (unconfounded) treatment.
}
\examples{
n <- 100; x <- matrix(rnorm(n*5),nrow=n)
a <- sample(3,n,replace=TRUE); y <- rnorm(n)

ate.res <- ate(y,a,x)

}
\references{
Robins JM, Rotnitzky A (1995). Semiparametric efficiency in multivariate regression models with missing data. \emph{Journal of the American Statistical Association}.

Hahn J (1998). On the role of the propensity score in efficient semiparametric estimation of average treatment effects. \emph{Econometrica}.

van der Laan MJ, Robins JM (2003). \emph{Unified Methods for Censored Longitudinal Data and Causality} (Springer).

Tsiatis AA (2006). \emph{Semiparametric Theory and Missing Data} (Springer).

Robins JM, Li L, Tchetgen Tchetgen ET, van der Vaart A (2008). Higher order influence functions and minimax estimation of nonlinear functionals. \emph{Probability and Statistics: Essays in Honor of David A. Freedman}.

Zheng W, van der Laan (2010). Asymptotic theory for cross-validated targeted maximum likelihood estimation \emph{UC Berkeley Division of Biostatistics Working Paper Series}.

Chernozhukov V, Chetverikov V, Demirer M, et al (2016). Double machine learning for treatment and causal parameters.
}
