% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att.R
\name{att}
\alias{att}
\title{Estimating average effect of treatment on the treated}
\usage{
att(y, a, x, nsplits=2, sl.lib=c("SL.earth","SL.gam","SL.glm","SL.glmnet",
  "SL.glm.interaction","SL.mean","SL.ranger"))
}
\arguments{
\item{y}{outcome of interest.}

\item{a}{binary treatment.}

\item{x}{covariate matrix.}

\item{nsplits}{integer number of sample splits for nuisance estimation.
If nsplits=1, sample splitting is not used, and nuisance functions are estimated
on full sample (in which case validity of SEs/CIs requires empirical
process conditions). Otherwise must have nsplits>1.}

\item{sl.lib}{algorithm library if using SuperLearner.
Default library includes "earth", "gam", "glm", "glmnet", "glm.interaction",
"mean", and "ranger".}
}
\value{
A list containing the following components:
\item{res}{ estimates/SEs/CIs/p-values for treated means and contrast.}
\item{nuis}{ subject-specific estimates of nuisance functions (i.e., propensity score and outcome regression) }
\item{ifvals}{ vector of estimated influence function values.}
}
\description{
\code{att} is used to estimate the difference in mean outcome among treated subjects had a binary (unconfounded) treatment been withheld.
}
\examples{
n <- 1000
x <- matrix(rnorm(n * 5), nrow = n)
a <- rbinom(n, 1, .3)
y <- rnorm(n)

att.res <- att(y, a, x)
}
\references{
(Also see references for function \code{ate})

Kennedy EH, Sjolander A, Small DS (2015). Semiparametric causal inference in matched cohort studies. \emph{Biometrika}.
}
