% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctseff.R
\name{ctseff}
\alias{ctseff}
\title{Estimating average effect curve for continuous treatment}
\usage{
ctseff(x, a, y, nsplits=2,
sl.lib=c("SL.earth","SL.gam","SL.glm","SL.glmnet", "SL.glm.interaction","SL.mean","SL.ranger"))
}
\arguments{
\item{x}{covariate matrix.}

\item{a}{continuous treatment.}

\item{y}{outcome of interest.}

\item{nsplits}{integer number of sample splits for nuisance estimation.
If nsplits=1, sample splitting is not used, and nuisance functions are estimated
on full sample (in which case validity of SEs/CIs requires empirical
process conditions). Otherwise must have nsplits>1.}

\item{sl.lib}{algorithm library if using SuperLearner.
Default library includes "earth", "gam", "glm", "glmnet", "glm.interaction",
"mean", and "ranger".}
}
\value{
A list containing the following components:
\item{res}{ estimates/SEs/CIs/p-values for population means and relevant contrasts.}
\item{ifvals}{ matrix of estimated influence function values.}
}
\description{
\code{ctseff} is used to estimate the mean outcome in a population had all subjects received given levels of a continuous (unconfounded) treatment.
}
\examples{
n <- 1000; x <- matrix(rnorm(n*4),nrow=n)
a <- runif(n); y <- rnorm(n)

ctseff.res <- ctseff(x,a,y)

}
\references{
Kennedy EH, Ma Z, McHugh MD, Small DS (2017). Nonparametric methods for doubly robust estimation of continuous treatment effects. \emph{Journal of the Royal Statistical Society, Series B}. \href{https://arxiv.org/abs/1507.00747}{arxiv:1507.00747}
}
