% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsi.R
\name{ipsi}
\alias{ipsi}
\title{Estimating effects of incremental propensity score interventions}
\usage{
ipsi(dat, x.trt, x.out, delta.seq, nsplits)
}
\arguments{
\item{dat}{dataframe (in long not wide form if longitudinal) with columns ‘time’, ‘id’, outcome ‘y’, treatment ‘a’.}

\item{x.trt}{covariate matrix for treatment regression.}

\item{x.out}{covariate matrix for outcome regression.}

\item{delta.seq}{sequence of delta values.}

\item{nsplits}{number of sample splits.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
\code{ipsi} is used to estimate effects of incremental propensity score interventions.
}
\references{
Kennedy EH. Nonparametric causal effects based on incremental propensity score interventions. \href{https://arxiv.org/abs/1704.00211}{arxiv:1704.00211}
}
