% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlate.R
\name{ivlate}
\alias{ivlate}
\title{Estimating complier average effect of binary treatment using binary instrument}
\usage{
ivlate(y, a, z, x, nsplits=2, sl.lib=c("SL.earth","SL.gam","SL.glm","SL.glmnet",
  "SL.glm.interaction", "SL.mean","SL.ranger","rpart"))
}
\arguments{
\item{y}{outcome of interest.}

\item{a}{binary treatment.}

\item{z}{binary instrument.}

\item{x}{covariate matrix.}

\item{nsplits}{integer number of sample splits for nuisance estimation.
If nsplits=1, sample splitting is not used, and nuisance functions are estimated
on full sample (in which case validity of SEs/CIs requires empirical
process conditions). Otherwise must have nsplits>1.}

\item{sl.lib}{algorithm library for SuperLearner.
Default library includes "earth", "gam", "glm", "glmnet", "glm.interaction",
"mean", "ranger", "rpart".}
}
\value{
A list containing the following components:
\item{res}{ estimates/SEs/CIs/p-values for local average treatment effect E(Y(a=1)-Y(a=0)|A(z=1)>A(z=0)), as well as IV strength and sharpness.}
\item{nuis}{ subject-specific estimates of nuisance functions (i.e., IV propensity score and treatment/outcome regressions) }
\item{ifvals}{ matrix of estimated influence function values.}
}
\description{
\code{ivlate} is used to estimate the mean outcome among compliers (i.e., those encouraged by the instrument) had all subjects received treatment versus control.
}
\examples{
n <- 100; x <- matrix(rnorm(n*5),nrow=n)
z <- rbinom(n,1,0.5); a <- rbinom(n,1,0.6*z+0.2)
y <- rnorm(n)

ivlate.res <- ivlate(y,a,z,x)

}
\references{
(Also see references for function \code{ate})

Angrist JD, Imbens GW, Rubin DB (1996). Identification of causal effects using instrumental variables. \emph{Journal of the American Statistical Association}.

Abadie A (2003). Semiparametric instrumental variable estimation of treatment response models. \emph{Journal of Econometrics}.

Kennedy EH, Balakrishnan S, G'Sell M (2017). Complier classification with sharp instrumental variables. \emph{Working Paper}.
}
