% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ctseff.R
\name{plot.ctseff}
\alias{plot.ctseff}
\title{Plot estimated average effect curve for continuous treatment}
\usage{
plot.ctseff(ctseff.res)
}
\arguments{
\item{ctseff.res}{output from \code{ctseff} fit.}
}
\value{
A plot of estimated effect curve with pointwise confidence intervals.
}
\description{
\code{plot.ctseff} is used to plot results from \code{ctseff} fit.
}
\examples{
n <- 500; x <- matrix(rnorm(n*5),nrow=n)
a <- runif(n); y <- a + rnorm(n,sd=.5)

ce.res <- ctseff(y,a,x, bw.seq=seq(.2,2,length.out=100))
plot.ctseff(ce.res)

}
\references{
Kennedy EH, Ma Z, McHugh MD, Small DS (2017). Nonparametric methods for doubly robust estimation of continuous treatment effects. \emph{Journal of the Royal Statistical Society, Series B}. \href{https://arxiv.org/abs/1507.00747}{arxiv:1507.00747}
}
