% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.ranger.R
\name{SL.ranger}
\alias{SL.ranger}
\title{Add Ranger wrapper for SuperLearner}
\usage{
SL.ranger(Y, X, newX, family, ...)
}
\arguments{
\item{Y}{outcome vector.}

\item{X}{covariate dataframe for training.}

\item{newX}{covariate dataframe for predictions.}

\item{family}{link function (currently only supports "gaussian" identity link).}
}
\value{
Predictions and fits from \code{ranger}.
}
\description{
\code{SL.ranger} is a wrapper for \code{SuperLearner} that adds the fast random forests method \code{ranger}.
}
\references{
Wright MN, Ziegler A (2016). ranger: A fast implementation of random forests for high dimensional data in C++ and R. \emph{Journal of Statistical Software}.
}
