% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ate.R
\name{ate}
\alias{ate}
\title{Estimating average effect of discrete treatment}
\usage{
ate(x, a, y, nsplits=2,
sl.lib=c("SL.earth","SL.gam","SL.glm","SL.glmnet", "SL.glm.interaction","SL.mean","SL.ranger"))
}
\arguments{
\item{x}{covariate matrix.}

\item{a}{discrete treatment.}

\item{y}{outcome of interest.}

\item{nsplits}{integer number of sample splits for nuisance estimation.
If nsplits=1, sample splitting is not used, and nuisance functions are estimated
on full sample (in which case validity of SEs/CIs requires empirical
process conditions). Otherwise must have nsplits>1.}

\item{sl.lib}{algorithm library if using SuperLearner.
Default library includes "earth", "gam", "glm", "glmnet", "glm.interaction",
"mean", and "ranger".}
}
\value{
A list containing the following components:
\item{res}{ estimates/SEs/CIs/p-values for population means and relevant contrasts.}
\item{ifvals}{ matrix of estimated influence function values.}
}
\description{
\code{ate} is used to estimate the mean outcome in a population had all subjects received given levels of a discrete (unconfounded) treatment.
}
\examples{
n <- 1000; x <- matrix(rnorm(n*4),nrow=n)
a <- sample(4,n,replace=TRUE); y <- rnorm(n)

ate.res <- ate(x,a,y)

}
\references{
Tsiatis AA. \emph{Semiparametric Theory and Missing Data} (Springer, 2006).
}
