% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctseff.R
\name{ctseff}
\alias{ctseff}
\title{Estimating average effect curve for continuous treatment}
\usage{
ctseff(y, a, x, bw.seq, sl.lib=c("SL.earth","SL.gam","SL.glm","SL.glmnet",
  "SL.glm.interaction","SL.mean","SL.ranger"))
}
\arguments{
\item{y}{outcome of interest.}

\item{a}{continuous treatment.}

\item{x}{covariate matrix.}

\item{bw.seq}{sequence of bandwidth values.}

\item{sl.lib}{algorithm library for SuperLearner.
Default library includes "earth", "gam", "glm", "glmnet", "glm.interaction",
"mean", and "ranger".}
}
\value{
A list containing the following components:
\item{res}{ estimates/SEs/CIs for population means.}
\item{bw.risk}{ estimated risk at sequence of bandwidth values.}
}
\description{
\code{ctseff} is used to estimate the mean outcomes in a population had all subjects received given levels of a continuous (unconfounded) treatment.
}
\examples{
n <- 500
x <- matrix(rnorm(n * 5), nrow = n)
a <- runif(n)
y <- a + rnorm(n, sd = .5)

ce.res <- ctseff(y, a, x, bw.seq = seq(.2, 2, length.out = 100))
plot.ctseff(ce.res)

# check that bandwidth choice is minimizer
plot(ce.res$bw.risk$bw, ce.res$bw.risk$risk)
}
\references{
Kennedy EH, Ma Z, McHugh MD, Small DS (2017). Nonparametric methods for doubly robust estimation of continuous treatment effects. \emph{Journal of the Royal Statistical Society, Series B}. \href{https://arxiv.org/abs/1507.00747}{arxiv:1507.00747}
}
